split activation, onboarding, setup tests into separate files
